Moved duplicate code to convert module.
