+++ src/convert.rs
@@ -28,3 +28,34 @@ macro_rules! impl_bidi_through_ref {
 		impl_through_ref!($trait<$b> for $a);
 	}
 }
+
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
+pub enum TryFromEgmCartesianSpeedError {
+	WrongNumberOfValues(usize),
+}
+
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
+pub enum TryFromEgmPoseError {
+	MissingPosition,
+	MissingOrientation,
+}
+
+impl std::fmt::Display for TryFromEgmCartesianSpeedError {
+	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+		match self {
+			Self::WrongNumberOfValues(x) => write!(f, "wrong number of values, expected 3, got {}", x),
+		}
+	}
+}
+
+impl std::fmt::Display for TryFromEgmPoseError {
+	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+		match self {
+			Self::MissingPosition => write!(f, "missing field: pos"),
+			Self::MissingOrientation => write!(f, "missing field: orient"),
+		}
+	}
+}
+
+impl std::error::Error for TryFromEgmCartesianSpeedError {}
+impl std::error::Error for TryFromEgmPoseError {}
