+++ src/glam.rs
@@ -1,5 +1,6 @@
 use crate::msg;
 
+use crate::convert::{TryFromEgmCartesianSpeedError, TryFromEgmPoseError};
 use std::convert::TryFrom;
 
 // Vector3
@@ -97,7 +98,7 @@ impl TryFrom<&msg::EgmPose> for glam::DAffine3 {
 
 		Ok(glam::DAffine3::from_rotation_translation(
             orientation.into(),
-			glam::DVec3::from(position).into(),
+			glam::DVec3::from(position),
 		))
 	}
 }
@@ -111,34 +112,3 @@ impl From<&glam::DAffine3> for msg::EgmPose {
 
 impl_through_ref!(From<glam::DAffine3> for msg::EgmPose);
 impl_through_ref!(TryFrom<msg::EgmPose> for glam::DAffine3);
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum TryFromEgmCartesianSpeedError {
-	WrongNumberOfValues(usize),
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum TryFromEgmPoseError {
-	MissingPosition,
-	MissingOrientation,
-}
-
-impl std::fmt::Display for TryFromEgmCartesianSpeedError {
-	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-		match self {
-			Self::WrongNumberOfValues(x) => write!(f, "wrong number of values, expected 3, got {}", x),
-		}
-	}
-}
-
-impl std::fmt::Display for TryFromEgmPoseError {
-	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-		match self {
-			Self::MissingPosition => write!(f, "missing field: pos"),
-			Self::MissingOrientation => write!(f, "missing field: orient"),
-		}
-	}
-}
-
-impl std::error::Error for TryFromEgmCartesianSpeedError {}
-impl std::error::Error for TryFromEgmPoseError {}
