+++ src/nalgebra.rs
@@ -1,5 +1,6 @@
 use crate::msg;
 
+use crate::convert::{TryFromEgmCartesianSpeedError, TryFromEgmPoseError};
 use std::convert::TryFrom;
 
 // Vector3
@@ -110,34 +111,3 @@ impl From<&nalgebra::Isometry3<f64>> for msg::EgmPose {
 
 impl_through_ref!(From<nalgebra::Isometry3<f64>> for msg::EgmPose);
 impl_through_ref!(TryFrom<msg::EgmPose> for nalgebra::Isometry3<f64>);
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum TryFromEgmCartesianSpeedError {
-	WrongNumberOfValues(usize),
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum TryFromEgmPoseError {
-	MissingPosition,
-	MissingOrientation,
-}
-
-impl std::fmt::Display for TryFromEgmCartesianSpeedError {
-	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-		match self {
-			Self::WrongNumberOfValues(x) => write!(f, "wrong number of values, expected 3, got {}", x),
-		}
-	}
-}
-
-impl std::fmt::Display for TryFromEgmPoseError {
-	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-		match self {
-			Self::MissingPosition => write!(f, "missing field: pos"),
-			Self::MissingOrientation => write!(f, "missing field: orient"),
-		}
-	}
-}
-
-impl std::error::Error for TryFromEgmCartesianSpeedError {}
-impl std::error::Error for TryFromEgmPoseError {}
